Extract most of the gradle check into a validator
