+++ script/tool/test/gradle_check_command_test.dart
@@ -7,6 +7,7 @@ import 'package:file/file.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_utils.dart';
 import 'package:flutter_plugin_tools/src/gradle_check_command.dart';
+import 'package:flutter_plugin_tools/src/validators/gradle_validator.dart';
 import 'package:git/git.dart';
 import 'package:test/test.dart';
 
@@ -165,7 +166,7 @@ gradle.projectsEvaluated {
     buildGradle.writeAsStringSync('''
 allprojects {
     repositories {
-        ${includeArtifactHub ? GradleCheckCommand.exampleRootGradleArtifactHubString : ''}
+        ${includeArtifactHub ? GradleValidator.exampleRootGradleArtifactHubString : ''}
         google()
         mavenCentral()
     }
@@ -226,7 +227,7 @@ pluginManagement {
     }
 }
 
-${includeArtifactDocumentation ? '// See ${GradleCheckCommand.artifactHubDocumentationString} for more info.' : ''}
+${includeArtifactDocumentation ? '// See ${GradleValidator.artifactHubDocumentationString} for more info.' : ''}
 plugins {
     id("dev.flutter.flutter-plugin-loader") version "1.0.0"
     id("com.android.application") version "8.11.1" apply false
@@ -989,8 +990,8 @@ flutter {
         expect(
           output,
           containsAllInOrder(<Matcher>[
-            contains(GradleCheckCommand.exampleRootGradleArtifactHubString),
-            contains(GradleCheckCommand.exampleSettingsArtifactHubString),
+            contains(GradleValidator.exampleRootGradleArtifactHubString),
+            contains(GradleValidator.exampleSettingsArtifactHubString),
           ]),
         );
       },
@@ -1025,7 +1026,7 @@ flutter {
       expect(
         output,
         containsAllInOrder(<Matcher>[
-          contains(GradleCheckCommand.exampleRootGradleArtifactHubString),
+          contains(GradleValidator.exampleRootGradleArtifactHubString),
         ]),
       );
     });
@@ -1059,12 +1060,12 @@ flutter {
       expect(
         output,
         containsAllInOrder(<Matcher>[
-          contains(GradleCheckCommand.exampleSettingsArtifactHubString),
+          contains(GradleValidator.exampleSettingsArtifactHubString),
         ]),
       );
       expect(
         output,
-        isNot(contains(GradleCheckCommand.exampleRootGradleArtifactHubString)),
+        isNot(contains(GradleValidator.exampleRootGradleArtifactHubString)),
       );
     });
 
@@ -1101,7 +1102,7 @@ flutter {
         expect(
           output,
           containsAllInOrder(<Matcher>[
-            contains(GradleCheckCommand.artifactHubDocumentationString),
+            contains(GradleValidator.artifactHubDocumentationString),
           ]),
         );
       },
